Move keyring logic into pkg/keyring package
Extracts SecureStore, ErrKeyNotFound, zalandoStore, fallbackStore,
fileStore, MemoryStore, NewSecureStore, and IsUsingInsecureStorage into
a new pkg/keyring package to keep credential-storage concerns separate
from CLI configuration. pkg/config now holds a thin wrapper that wires
the keyring to its config-directory path.

Committed-By-Agent: claude
