+++ pkg/login/interactive_login_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/stripe/stripe-cli/pkg/config"
+	"github.com/stripe/stripe-cli/pkg/keyring"
 	"github.com/stripe/stripe-cli/pkg/login/acct"
 )
 
@@ -30,7 +31,7 @@ func setupInteractiveLoginConfig(t *testing.T) (*config.Config, func()) {
 		ProfilesFile: profilesFile,
 	}
 	c.InitConfig()
-	config.KeyRing = config.NewMemoryStore(nil)
+	config.KeyRing = keyring.NewMemoryStore(nil)
 	cleanup := func() {
 		os.Remove(profilesFile)
 	}
